move: 상속 관계 파일 이동 #7
상속 관계의 파일을 묶음
