+++ BE/src/main/java/com/healthyreal/be/domain/OAuth2UserInfo/OAuth2UserInfo.java
@@ -1,16 +1,14 @@
-package com.healthyreal.be.domain;
+package com.healthyreal.be.domain.OAuth2UserInfo;
 
 import java.util.Map;
+import lombok.Getter;
 
+@Getter
 public abstract class OAuth2UserInfo {
 	protected Map<String, Object> attributes;
 
 	public OAuth2UserInfo(Map<String, Object> attributes) {
 		this.attributes = attributes;
-	}
-
-	public Map<String, Object> getAttributes() {
-		return attributes;
 	}
 
 	public abstract String getId();
